LocalAppiumServer: Move class under Maze namespace, adding docs and extra checks
